Renamed testcase.cpp to testcase_reader.cpp.
