+++ tests/CMakeLists.txt
@@ -8,7 +8,7 @@ target_link_libraries(link-t predict m)
 add_test(NAME link COMMAND link-t)
 
 #test against satellite data
-add_executable(orbit-t orbit-t.cpp testcase.cpp)
+add_executable(orbit-t orbit-t.cpp testcase_reader.cpp)
 target_link_libraries(orbit-t predict)
 file(GLOB files "${LIBPREDICT_TEST_DATA_DIR}/testcase_satellite_*")
 foreach(file ${files})
@@ -16,7 +16,7 @@ foreach(file ${files})
 endforeach()
 
 #test aos/los on satellite data
-add_executable(aoslos-t aoslos-t.cpp testcase.cpp)
+add_executable(aoslos-t aoslos-t.cpp testcase_reader.cpp)
 target_link_libraries(aoslos-t predict)
 file(GLOB files "${LIBPREDICT_TEST_DATA_DIR}/testcase_satellite_*")
 foreach(file ${files})
@@ -24,7 +24,7 @@ foreach(file ${files})
 endforeach()
 
 #test against sun data
-add_executable(sun-t sun-t.cpp testcase.cpp)
+add_executable(sun-t sun-t.cpp testcase_reader.cpp)
 target_link_libraries(sun-t predict)
 file(GLOB files "${LIBPREDICT_TEST_DATA_DIR}/testcase_sun_*")
 foreach(file ${files})
@@ -32,7 +32,7 @@ foreach(file ${files})
 endforeach()
 
 #test against moon data
-add_executable(moon-t moon-t.cpp testcase.cpp)
+add_executable(moon-t moon-t.cpp testcase_reader.cpp)
 target_link_libraries(moon-t predict)
 file(GLOB files "${LIBPREDICT_TEST_DATA_DIR}/testcase_moon_*")
 foreach(file ${files})
