+++ tests/aoslos-t.cpp
@@ -4,7 +4,7 @@
 #include <unistd.h>
 #include <stdlib.h>
 #include <predict/predict.h>
-#include "testcase.h"
+#include "testcase_reader.h"
 #include <iostream>
 using namespace std;
