+++ tests/orbit-t.cpp
@@ -5,7 +5,7 @@
 #include <stdlib.h>
 #include <predict/predict.h>
 
-#include "testcase.h"
+#include "testcase_reader.h"
 
 #include <iostream>
 using namespace std;
